study_tornado/study_school/requirements.txt
tornado
python-memcached
peewee
simplejson
Pillow
marshmallow
PyJWT
xmltodict